+++ packages/smoldot/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Added optional `SmoldotOptions` parameter to `startFromWorker`
+
 ## 0.0.1 - 2024-04-10
 
 ### Changed